converter :bug: type fix
