+++ core/src/main/java/com/sequenceiq/cloudbreak/converter/v4/stacks/cluster/ambari/AmbariRepoDetailsToAmbariRepositoryV4ResponseConverter.java
@@ -3,19 +3,18 @@ package com.sequenceiq.cloudbreak.converter.v4.stacks.cluster.ambari;
 import org.springframework.stereotype.Component;
 
 import com.sequenceiq.cloudbreak.api.endpoint.v4.stacks.response.cluster.ambari.ambarirepository.AmbariRepositoryV4Response;
-import com.sequenceiq.cloudbreak.cloud.model.AmbariRepo;
+import com.sequenceiq.cloudbreak.cloud.model.component.AmbariRepoDetails;
 import com.sequenceiq.cloudbreak.converter.AbstractConversionServiceAwareConverter;
 
 @Component
 public class AmbariRepoDetailsToAmbariRepositoryV4ResponseConverter
-        extends AbstractConversionServiceAwareConverter<AmbariRepo, AmbariRepositoryV4Response> {
+        extends AbstractConversionServiceAwareConverter<AmbariRepoDetails, AmbariRepositoryV4Response> {
 
     @Override
-    public AmbariRepositoryV4Response convert(AmbariRepo source) {
+    public AmbariRepositoryV4Response convert(AmbariRepoDetails source) {
         AmbariRepositoryV4Response ambariRepoDetailsJson = new AmbariRepositoryV4Response();
-        ambariRepoDetailsJson.setBaseUrl(source.getBaseUrl());
-        ambariRepoDetailsJson.setGpgKeyUrl(source.getGpgKeyUrl());
-        ambariRepoDetailsJson.setVersion(source.getVersion());
+        ambariRepoDetailsJson.setBaseUrl(source.getBaseurl());
+        ambariRepoDetailsJson.setGpgKeyUrl(source.getGpgkey());
         return ambariRepoDetailsJson;
     }
