+++ core/src/main/java/com/sequenceiq/cloudbreak/converter/v4/stacks/cluster/ambari/AmbariRepoToAmbariRepositoryV4ResponseConverter.java
@@ -1,21 +1,18 @@
 package com.sequenceiq.cloudbreak.converter.v4.stacks.cluster.ambari;
 
-import org.springframework.stereotype.Component;
-
 import com.sequenceiq.cloudbreak.api.endpoint.v4.stacks.response.cluster.ambari.ambarirepository.AmbariRepositoryV4Response;
 import com.sequenceiq.cloudbreak.cloud.model.AmbariRepo;
 import com.sequenceiq.cloudbreak.converter.AbstractConversionServiceAwareConverter;
 
-@Component
-public class AmbariRepoToAmbariRepositoryV4ResponseConverter extends AbstractConversionServiceAwareConverter<AmbariRepositoryV4Response, AmbariRepo> {
+public class AmbariRepoToAmbariRepositoryV4ResponseConverter extends AbstractConversionServiceAwareConverter<AmbariRepo, AmbariRepositoryV4Response> {
 
     @Override
-    public AmbariRepo convert(AmbariRepositoryV4Response source) {
-        AmbariRepo ambariRepo = new AmbariRepo();
-        ambariRepo.setPredefined(Boolean.FALSE);
-        ambariRepo.setVersion(source.getVersion());
-        ambariRepo.setBaseUrl(source.getBaseUrl());
-        ambariRepo.setGpgKeyUrl(source.getGpgKeyUrl());
-        return ambariRepo;
+    public AmbariRepositoryV4Response convert(AmbariRepo source) {
+        var response = new AmbariRepositoryV4Response();
+        response.setBaseUrl(source.getBaseUrl());
+        response.setGpgKeyUrl(source.getGpgKeyUrl());
+        response.setVersion(source.getVersion());
+        return response;
     }
+
 }
